rethink author link to increase hitbot
