+++ site/app/_components/AuthorLink.tsx
@@ -1,18 +1,44 @@
+import clsx from 'clsx';
 import React from 'react';
+
+interface AuthorImageProps {
+  handle: string;
+  className?: string;
+}
+
+const AuthorImage: React.FC<AuthorImageProps> = ({ handle, className }) => {
+  const username = handle.replace('@', '');
+  const githubUrl = `https://github.com/${username}`;
+
+  // eslint-disable-next-line @next/next/no-img-element
+  return <img src={`${githubUrl}.png?size=100`} alt={`Avatar for ${handle}`} className={className} />;
+};
 
 interface AuthorLinkProps {
   handle: string;
   className?: string;
-  children?: React.ReactNode;
 }
 
-const AuthorLink: React.FC<AuthorLinkProps> = ({ handle, className, children }) => {
+const AuthorLink: React.FC<AuthorLinkProps> = ({ handle, className }) => {
   const username = handle.replace('@', '');
   const githubUrl = `https://github.com/${username}`;
 
   return (
-    <a href={githubUrl} className={className} target="_blank" rel="noopener noreferrer">
-      {children}
+    <a
+      href={githubUrl}
+      className={clsx('flex gap-0.5 flex-row items-center mb-1 group', className)}
+      target="_blank"
+      rel="noopener noreferrer"
+    >
+      <span className="rounded-1 overflow-clip">
+        <AuthorImage handle={handle} className="w-2 h-2" />
+      </span>
+      <span className="font-mono leading-mono font-normal">
+        By{' '}
+        <span className="underline underline-offset-mono decoration-link group-hover:no-underline group-focus-visible:no-underline">
+          {handle}
+        </span>
+      </span>
     </a>
   );
 };
